Improved Serializer
commit c55926cb6af51f703ee7e06b6ea810dd19085e61
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 18:16:54 2022 +0900

    wip

commit 3bb2481283431b1a82c9161f1af44bdfd0d5151e
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 16:20:49 2022 +0900

    wip

commit a8cf60563ba56123d3c6a3bdccd0f3e6719df0f5
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 15:08:10 2022 +0900

    wip

commit 4ce0b257cccea2c8fdba9977ad9bd940f8677f3c
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 11:45:39 2022 +0900

    wip

commit 88e76d3d8cdc91c7314fe1dffdfad2310931142b
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 11:26:03 2022 +0900

    wip

commit f979ce8d625f7c75007133b502f92f0b1a96788c
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 10:58:39 2022 +0900

    wip

commit 6afeac1d66e0e3cdd8075f21e70b89af1e9d3546
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 10:45:53 2022 +0900

    wip

commit 642e5e2a6e2146ea5ef13b43a0eba8f7451e7c3a
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 22 10:21:34 2022 +0900

    wip

commit fd4d6359d942ef504de3f376fc10b2828e06ef9e
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 18:19:14 2022 +0900

    wip

commit 39919132b2a3b6e25aab83e6153455140fa73554
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 17:09:14 2022 +0900

    wip

commit 04c47a234905b1591c8b2bb25cdd72e77d1a57e0
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 14:28:35 2022 +0900

    wip

commit eb6af4d4878826fced40d9704e34d64806711e5f
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 14:02:27 2022 +0900

    wip

commit dbeb67e46be16718c1555281b48fe45b25c9822c
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 13:31:19 2022 +0900

    wip

commit a6214336abffe8e3e09593bd6b5688d31680a208
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 12:32:53 2022 +0900

    wip

commit b8c8de1c35d68e746333b8b388de8d428494ed8c
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 11:59:16 2022 +0900

    wip

commit df094987cd9f1481175e0714b2a5ec2870a002bf
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 11:17:14 2022 +0900

    wip

commit 8f9daa0034a16f0a329ccbcc5fef42392ac26113
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 10:44:18 2022 +0900

    wip

commit 99adab4a856388d8110f3a1fe957010142fb886d
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 20 10:02:41 2022 +0900

    wip

commit 67f8154ed3da29be6a626f86e8abab2e5ce191b9
Author: lriki <lriki.net@gmail.com>
Date:   Tue Jul 19 18:31:48 2022 +0900

    wip

commit 60cd4669aae1d082a4db4435a51654e7273d649d
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 15 19:51:49 2022 +0900

    wip

commit 642d7943e4f033ae2d150ce61f26be35bafb6804
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 15 19:42:57 2022 +0900

    wip

commit 5ac259601ff0bcdd17421b13b952235225fe3165
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 15 18:42:43 2022 +0900

    wip

+++ lumino/Bitmap/src/Bitmap.cpp
@@ -387,7 +387,7 @@ void Bitmap2D::flipVerticalFlow()
 	if (LN_REQUIRE(m_format != PixelFormat::Unknown)) return;
 	//if (LN_REQUIRE(m_format != PixelFormat::A1)) return;
 
-	int pixelSize = detail::BlitHelper::getPixelSize(m_format);
+	size_t pixelSize = detail::BlitHelper::getPixelSize(m_format);
 	if (pixelSize == 1)
 	{
 		// XOR で工夫すると演算回数が少なくなるとか最適化の余地はあるけど、
